+++ UI/admin/src/views/Products/AllProduct.js
@@ -1,10 +1,13 @@
 import React, { useEffect, useState } from "react";
 import Base from "../../Components/Base";
-import { getAllProducts, deleteProduct } from "../../api/helper";
+import { getAllProducts, updateProduct, deleteProduct } from "../../api/helper";
 import Table from "../../Components/Table";
 import Avatar from "@material-ui/core/Avatar";
 import { getMediaUrl } from "../../api/functions";
 import { useHistory } from "react-router-dom";
+import { toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
+import swal from "sweetalert2";
 
 function Index() {
   const history = useHistory();
@@ -29,14 +32,51 @@ function Index() {
         Header: "English Name",
         accessor: "eng_name", // accessor is the "key" in the data
       },
+
       {
         Header: "Nepali Name",
         accessor: "nep_name", // accessor is the "key" in the data
       },
       {
+        Header: "status",
+        accessor: "status",
+        Cell: (value) => (
+          <span>
+            <a
+              style={{ cursor: "pointer" }}
+              className={
+                "text-" +
+                (value && value.value == "Active" ? "success" : "secondary")
+              }
+              onClick={() => toggleStatus(value.data[value.row.index])}
+            >
+              {value.value}
+            </a>
+          </span>
+        ),
+      },
+      {
         Header: "Category",
         accessor: "category",
         Cell: (value) => <span>{value.value.name}</span>,
+      },
+      {
+        Header: "Featured",
+        accessor: "featured",
+        Cell: (value) => (
+          <span>
+            <a
+              style={{ cursor: "pointer" }}
+              className={
+                "text-" +
+                ((value && value.value) == true ? "success" : "danger")
+              }
+              onClick={() => toggleFeatured(value.data[value.row.index])}
+            >
+              {value && value.value ? <i className="fa fa-check"></i> : <i className="fa fa-times"></i>}
+            </a>
+          </span>
+        ),
       },
 
       {
@@ -55,6 +95,39 @@ function Index() {
     ],
     []
   );
+  const toggleFeatured = (data) => {
+    console.log(data);
+    let toSet=data.featured === true ? false : true 
+    updateProduct(
+      { _id: data._id },
+      { featured: toSet}
+    )
+      .then((resp) => {
+        fetchProducts();
+        swal.fire("Product is now " + (toSet?'featured':'normal'));
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+  const toggleStatus = (data) => {
+    console.log(data);
+    let toSet = data.status && data.status === "Active" ? "Inactive" : "Active";
+    updateProduct(
+      { _id: data._id },
+      {
+        status: toSet,
+      }
+    )
+      .then((resp) => {
+        swal.fire("Product is now " + toSet);
+
+        fetchProducts();
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
   const Actions = ({ values }) => {
     return (
       <div className="btn-group rounded-0">
@@ -74,15 +147,32 @@ function Index() {
     );
   };
   const deleteRow = (product) => {
-    deleteProduct(product)
-      .then((resp) => {
-        fetchProducts();
+    swal
+      .fire({
+        title: "Do you want to delete this product ?",
+        showCancelButton: true,
+        confirmButtonText: `yes delete`,
       })
-      .catch((err) => {
-        console.log(err);
+      .then((willDelete) => {
+        if (willDelete.isConfirmed) {
+          deleteProduct(product)
+            .then((resp) => {
+              fetchProducts();
+              swal.fire("Product Deleted");
+            })
+            .catch((err) => {
+              console.log(err);
+            });
+        } else {
+          swal.fire("Deletion cancelled!");
+        }
       });
   };
   const editRow = (product) => {
+    swal.fire({
+      title: "Sorry This feature is under maintainance",
+      icon: "info",
+    });
     //  history.push("/product/edit/" + product._id);
   };
   return (
